requirements.txt
tensorflow==2.10.0
keras==2.10.0
numpy
pandas
matplotlib
scikit-learn
opencv-python
jupyter
pyserial